First draft with time setting functionalities
This code updates a series of 6 LEDs located in a row representing a 6 bit "byte" once per second. It blinks once for hours and twice for minutes, hence giving the hour in a binary clock fashion with a minimum of elements.
The 6 bit display of information allows to produce a representation of numbers from 1 to 63, which is just enough to represent up to 60 seconds in a minute or 24 hours in a day.